+++ src/components/index.js
@@ -1,2 +1,3 @@
 export * from './MarcAuthoritiesAppContext';
 export * from './KeyShortCutsWrapper';
+export * from './DateRangeFieldset';