+++ src/views/AuthoritiesSearch/ReportsMenu/index.js
@@ -1,0 +1,1 @@
+export { default as ReportsMenu } from './ReportsMenu';