+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [UIMARCAUTH-224](https://issues.folio.org/browse/UIMARCAUTH-224) Remove unneeded `react-redux`
 - [UIMARCAUTH-221](https://issues.folio.org/browse/UIMARCAUTH-221) Align the module with mod-search API breaking change
 - [UIMARCAUTH-226](https://issues.folio.org/browse/UIMARCAUTH-226) Bump quick-marc to 6.0.0
+- [UIMARCAUTH-208](https://issues.folio.org/browse/UIMARCAUTH-208) Global Heading Report | UX | MARC authority headings updates report
 
 ## [2.0.1](https://github.com/folio-org/ui-marc-authorities/tree/v2.0.1) (2022-11-28)